requirements.txt
# Core
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.43.0

# Modeling
scikit-learn==1.4.2
scipy==1.13.0
statsmodels==0.14.1
shap==0.45.0

# Visualization
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0

# Data / storage
python-dotenv==1.0.1
pyarrow==15.0.2

# Dev
pytest==8.1.1
black==24.3.0
ruff==0.3.5
